Style changes
The code being so poorly formatted is driving me up the wall

- format code
- turn webview into an interface
- webview native into a class
- use enhanced switch
