+++ avaje-webview/src/main/java/io/avaje/webview/DWebView.java
@@ -1,0 +1,342 @@
+/**
+ * MIT LICENSE
+ *
+ * <p>Copyright (c) 2024 Alex Bowles @ Casterlabs
+ *
+ * <p>Permission is hereby granted, free of charge, to any person obtaining a copy of this software
+ * and associated documentation files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute,
+ * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * <p>The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
+ *
+ * <p>THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING
+ * BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package io.avaje.webview;
+
+import static io.avaje.webview.platform.OSFamily.WINDOWS;
+import static io.avaje.webview.platform.Platform.OS_FAMILY;
+import static java.lang.System.Logger.Level.DEBUG;
+
+import module java.base;
+
+import module org.jspecify;
+
+import io.avaje.webview.WebviewNative.BindCallback;
+import io.avaje.webview.WebviewNative.DispatchCallback;
+
+/**
+ * Webview browser window.
+ *
+ * <pre>{@code
+ * Webview wv = Webview.builder()
+ *          .debug(true)
+ *          .title("My App")
+ *          .width(1000)
+ *          .height(800)
+ *          .url("http://localhost:" + port)
+ *          .build();
+ *
+ *  wv.run(); // Run the webview event loop, the webview is fully disposed when this returns.
+ *  wv.close(); // Free any resources allocated.
+ *
+ * }</pre>
+ */
+public final class DWebView implements Webview {
+
+  private static final System.Logger log = System.getLogger("io.avaje.webview");
+
+  private static final int WV_HINT_NONE = 0;
+  private static final int WV_HINT_MIN = 1;
+  private static final int WV_HINT_MAX = 2;
+  private static final int WV_HINT_FIXED = 3;
+
+  private final MemorySegment webview;
+  private final WebviewNative N;
+  private final Arena arena = Arena.ofAuto();
+
+  public static WebviewBuilder builder() {
+    return new WebviewBuilder();
+  }
+
+  DWebView(
+      WebviewNative n,
+      boolean debug,
+      @Nullable MemorySegment windowPointer,
+      int width,
+      int height) {
+    N = n;
+    this.webview =
+        n.webview_create(debug, windowPointer == null ? MemorySegment.NULL : windowPointer);
+    this.setSize(width, height);
+  }
+
+  /** Use this only if you absolutely know what you're doing. */
+  @Override
+  public MemorySegment nativeWindowPointer() {
+    return N.webview_get_window(webview);
+  }
+
+  @Override
+  public void setHTML(@Nullable String html) {
+    N.webview_set_html(webview, html);
+  }
+
+  @Override
+  public void loadURL(@Nullable String url) {
+    if (url == null) {
+      url = "about:blank";
+    }
+    N.webview_navigate(webview, url);
+  }
+
+  @Override
+  public void setTitle(@NonNull String title) {
+    N.webview_set_title(webview, title);
+  }
+
+  @Override
+  public void setMinSize(int width, int height) {
+    N.webview_set_size(webview, width, height, WV_HINT_MIN);
+  }
+
+  @Override
+  public void setMaxSize(int width, int height) {
+    N.webview_set_size(webview, width, height, WV_HINT_MAX);
+  }
+
+  @Override
+  public void setSize(int width, int height) {
+    N.webview_set_size(webview, width, height, WV_HINT_NONE);
+  }
+
+  @Override
+  public void setFixedSize(int width, int height) {
+    N.webview_set_size(webview, width, height, WV_HINT_FIXED);
+  }
+
+  /**
+   * Sets the script to be run on page load. Defaults to no nested access (false).
+   *
+   * @implNote This get's called AFTER window.load.
+   * @param script
+   * @see #setInitScript(String, boolean)
+   */
+  @Override
+  public void setInitScript(@NonNull String script) {
+    this.setInitScript(script, false);
+  }
+
+  /**
+   * Sets the script to be run on page load.
+   *
+   * @implNote This get's called AFTER window.load.
+   * @param script
+   * @param allowNestedAccess whether or not to inject the script into nested iframes.
+   */
+  @Override
+  public void setInitScript(@NonNull String script, boolean allowNestedAccess) {
+    script =
+        String.format(
+            """
+      	(() => {
+      	try {
+      	if (window.top == window.self || %b) {
+      	%s
+      	}
+      	} catch (e) {
+      	console.error('[Webview]', 'An error occurred whilst evaluating init script:', %s, e);
+      	}
+      	})();""",
+            allowNestedAccess, script, '"' + WebviewUtil.jsonEscape(script) + '"');
+
+    N.webview_init(webview, script);
+  }
+
+  /**
+   * Executes the given script NOW.
+   *
+   * @param script
+   */
+  @Override
+  public void eval(@NonNull String script) {
+    this.dispatch(
+        () -> {
+          N.webview_eval(
+              webview,
+              String.format(
+                  """
+        	try {
+        	%s
+        	} catch (e) {
+        	console.error('[Webview]', 'An error occurred whilst evaluating script:', %s, e);
+        	}""",
+                  script, '"' + WebviewUtil.jsonEscape(script) + '"'));
+        });
+  }
+
+  /**
+   * Binds a function to the JavaScript environment on page load.
+   *
+   * @implNote This get's called AFTER window.load.
+   * @implSpec After calling the function in JavaScript you will get a Promise instead of the value.
+   *     This is to prevent you from locking up the browser while waiting on your Java code to
+   *     execute and generate a return value.
+   * @param name The name to be used for the function, e.g "foo" to get foo().
+   * @param handler The callback handler, accepts a JsonArray (which are all arguments passed to the
+   *     function()) and returns a value which is of type JsonElement (can be null). Exceptions are
+   *     automatically passed back to JavaScript.
+   */
+  @Override
+  public void bind(@NonNull String name, @NonNull WebviewBindCallback handler) {
+    BindCallback callback =
+        (seq, req, _) -> {
+          try {
+            req = WebviewUtil.forceSafeChars(req);
+
+            String result = handler.apply(req);
+            if (result == null) {
+              result = "null";
+            }
+
+            N.webview_return(webview, seq, false, WebviewUtil.forceSafeChars(result));
+          } catch (Throwable e) {
+            e.printStackTrace();
+
+            String exceptionJson =
+                '"' + WebviewUtil.jsonEscape(WebviewUtil.getExceptionStack(e)) + '"';
+
+            N.webview_return(webview, seq, true, exceptionJson);
+          }
+        };
+
+    // Create upcall stub for the callback
+    MemorySegment callbackStub =
+        Linker.nativeLinker()
+            .upcallStub(createBindCallbackHandle(callback), BindCallback.DESCRIPTOR, arena);
+
+    N.webview_bind(webview, name, callbackStub, 0);
+  }
+
+  private static MethodHandle createBindCallbackHandle(BindCallback callback) {
+    try {
+      return MethodHandles.lookup()
+          .bind(
+              callback,
+              "actualCallBack",
+              MethodType.methodType(void.class, long.class, MemorySegment.class, long.class))
+          .asType(MethodType.methodType(void.class, long.class, MemorySegment.class, long.class));
+    } catch (Exception e) {
+      throw new RuntimeException("Failed to create callback handle", e);
+    }
+  }
+
+  /**
+   * Unbinds a function, removing it from future pages.
+   *
+   * @param name The name of the function.
+   */
+  @Override
+  public void unbind(@NonNull String name) {
+    N.webview_unbind(webview, name);
+  }
+
+  /**
+   * Executes an event on the event thread.
+   *
+   * @implNote Use this only if you absolutely know what you're doing.
+   */
+  @Override
+  public void dispatch(@NonNull Runnable handler) {
+
+    // Create upcall stub for the dispatch callback
+    MemorySegment callbackStub =
+        Linker.nativeLinker()
+            .upcallStub(
+                createDispatchCallbackHandle((_, _) -> handler.run()),
+                DispatchCallback.DESCRIPTOR,
+                arena);
+
+    N.webview_dispatch(webview, callbackStub, 0);
+  }
+
+  private static MethodHandle createDispatchCallbackHandle(DispatchCallback callback) {
+    try {
+      return MethodHandles.lookup()
+          .bind(
+              callback,
+              "callback",
+              MethodType.methodType(void.class, MemorySegment.class, long.class));
+    } catch (Exception e) {
+      throw new RuntimeException("Failed to create callback handle", e);
+    }
+  }
+
+  /**
+   * Executes the webview event loop until the user presses "X" on the window.
+   *
+   * @see #close()
+   */
+  @Override
+  public void run() {
+    N.webview_run(webview);
+    log.log(DEBUG, "destroy and terminate");
+    N.webview_destroy(webview);
+    N.webview_terminate(webview);
+  }
+
+  /**
+   * Executes the webview event loop asynchronously until the user presses "X" on the window.
+   *
+   * @see #close()
+   */
+  @Override
+  public void runAsync() {
+    Thread t = new Thread(this);
+    t.setDaemon(false);
+    t.setName("Webview RunAsync Thread - #" + this.hashCode());
+    t.start();
+  }
+
+  /** Closes the webview, call this to end the event loop and free up resources. */
+  @Override
+  public void close() {
+    log.log(DEBUG, "close");
+    N.webview_terminate(webview);
+  }
+
+  @Override
+  public void setDarkAppearance(boolean shouldAppearDark) {
+    if (WINDOWS.equals(OS_FAMILY)) {
+      WindowsHelper.setWindowAppearance(this, shouldAppearDark);
+    }
+  }
+
+  @Override
+  public Webview maximizeWindow() {
+    if (WINDOWS.equals(OS_FAMILY)) {
+      WindowsHelper.maximizeWindow(this);
+    }
+    return this;
+  }
+
+  @Override
+  public Webview fullscreen() {
+    if (WINDOWS.equals(OS_FAMILY)) {
+      WindowsHelper.fullscreen(this);
+    }
+    return this;
+  }
+
+  @Override
+  public String version() {
+    var versionInfo = N.webview_version();
+    return versionInfo.versionNumber();
+  }
+}
