+++ avaje-webview/src/main/java/io/avaje/webview/Webview.java
@@ -21,89 +21,31 @@
 package io.avaje.webview;
 
 import module java.base;
+
 import module org.jspecify;
 
-import static io.avaje.webview.WebviewNative.*;
-import static io.avaje.webview.platform.OSFamily.WINDOWS;
-import static io.avaje.webview.platform.Platform.OS_FAMILY;
-import static java.lang.System.Logger.Level.DEBUG;
+public interface Webview extends Closeable, Runnable {
 
-/**
- * Webview browser window.
- *
- * <pre>{@code
- * Webview wv = Webview.builder()
- *          .debug(true)
- *          .title("My App")
- *          .width(1000)
- *          .height(800)
- *          .url("http://localhost:" + port)
- *          .build();
- *
- *  wv.run(); // Run the webview event loop, the webview is fully disposed when this returns.
- *  wv.close(); // Free any resources allocated.
- *
- * }</pre>
- */
-public final class Webview implements Closeable, Runnable {
-
-  static final System.Logger log = System.getLogger("io.avaje.webview");
-
-  private final WebviewNative N;
-  private final MemorySegment webview;
-  private final Arena arena = Arena.ofAuto();
-
-  public static WebviewBuilder builder() {
+  static WebviewBuilder builder() {
     return new WebviewBuilder();
   }
 
-  Webview(
-      WebviewNative n,
-      boolean debug,
-      @Nullable MemorySegment windowPointer,
-      int width,
-      int height) {
-    this.N = n;
-    this.webview =
-        N.webview_create(debug, windowPointer == null ? MemorySegment.NULL : windowPointer);
-    this.setSize(width, height);
-  }
-
   /** Use this only if you absolutely know what you're doing. */
-  public MemorySegment nativeWindowPointer() {
-    return N.webview_get_window(webview);
-  }
+  MemorySegment nativeWindowPointer();
 
-  public void setHTML(@Nullable String html) {
-    N.webview_set_html(webview, html);
-  }
+  void setHTML(@Nullable String html);
 
-  public void loadURL(@Nullable String url) {
-    if (url == null) {
-      url = "about:blank";
-    }
-    N.webview_navigate(webview, url);
-  }
+  void loadURL(@Nullable String url);
 
-  public void setTitle(@NonNull String title) {
-    N.webview_set_title(webview, title);
-  }
+  void setTitle(@NonNull String title);
 
-  public void setMinSize(int width, int height) {
-    N.webview_set_size(webview, width, height, WV_HINT_MIN);
-  }
+  void setMinSize(int width, int height);
 
-  public void setMaxSize(int width, int height) {
-    N.webview_set_size(webview, width, height, WV_HINT_MAX);
-  }
+  void setMaxSize(int width, int height);
 
-  public void setSize(int width, int height) {
-    N.webview_set_size(webview, width, height, WV_HINT_NONE);
-  }
+  void setSize(int width, int height);
 
-  public void setFixedSize(int width, int height) {
-    N.webview_set_size(webview, width, height, WV_HINT_FIXED);
-  }
+  void setFixedSize(int width, int height);
 
   /**
    * Sets the script to be run on page load. Defaults to no nested access (false).
@@ -112,9 +54,7 @@ public final class Webview implements Closeable, Runnable {
    * @param script
    * @see #setInitScript(String, boolean)
    */
-  public void setInitScript(@NonNull String script) {
-    this.setInitScript(script, false);
-  }
+  void setInitScript(@NonNull String script);
 
   /**
    * Sets the script to be run on page load.
@@ -123,44 +63,14 @@ public final class Webview implements Closeable, Runnable {
    * @param script
    * @param allowNestedAccess whether or not to inject the script into nested iframes.
    */
-  public void setInitScript(@NonNull String script, boolean allowNestedAccess) {
-    script =
-        String.format(
-            """
-      	(() => {
-      	try {
-      	if (window.top == window.self || %b) {
-      	%s
-      	}
-      	} catch (e) {
-      	console.error('[Webview]', 'An error occurred whilst evaluating init script:', %s, e);
-      	}
-      	})();""",
-            allowNestedAccess, script, '"' + _WebviewUtil.jsonEscape(script) + '"');
-
-    N.webview_init(webview, script);
-  }
+  void setInitScript(@NonNull String script, boolean allowNestedAccess);
 
   /**
    * Executes the given script NOW.
    *
    * @param script
    */
-  public void eval(@NonNull String script) {
-    this.dispatch(
-        () -> {
-          N.webview_eval(
-              webview,
-              String.format(
-                  """
-        	try {
-        	%s
-        	} catch (e) {
-        	console.error('[Webview]', 'An error occurred whilst evaluating script:', %s, e);
-        	}""",
-                  script, '"' + _WebviewUtil.jsonEscape(script) + '"'));
-        });
-  }
+  void eval(@NonNull String script);
 
   /**
    * Binds a function to the JavaScript environment on page load.
@@ -174,85 +84,21 @@ public final class Webview implements Closeable, Runnable {
    *     function()) and returns a value which is of type JsonElement (can be null). Exceptions are
    *     automatically passed back to JavaScript.
    */
-  public void bind(@NonNull String name, @NonNull WebviewBindCallback handler) {
-    BindCallback callback =
-        (seq, req, arg) -> {
-          try {
-            req = _WebviewUtil.forceSafeChars(req);
-
-            String result = handler.apply(req);
-            if (result == null) {
-              result = "null";
-            }
-
-            N.webview_return(webview, seq, false, _WebviewUtil.forceSafeChars(result));
-          } catch (Throwable e) {
-            e.printStackTrace();
-
-            String exceptionJson =
-                '"' + _WebviewUtil.jsonEscape(_WebviewUtil.getExceptionStack(e)) + '"';
-
-            N.webview_return(webview, seq, true, exceptionJson);
-          }
-        };
-
-    // Create upcall stub for the callback
-    MemorySegment callbackStub =
-        Linker.nativeLinker()
-            .upcallStub(createBindCallbackHandle(callback), BindCallback.DESCRIPTOR, arena);
-
-    N.webview_bind(webview, name, callbackStub, 0);
-  }
-
-  private static MethodHandle createBindCallbackHandle(BindCallback callback) {
-    try {
-      return MethodHandles.lookup()
-          .bind(
-              callback,
-              "actualCallBack",
-              MethodType.methodType(void.class, long.class, MemorySegment.class, long.class))
-          .asType(MethodType.methodType(void.class, long.class, MemorySegment.class, long.class));
-    } catch (Exception e) {
-      throw new RuntimeException("Failed to create callback handle", e);
-    }
-  }
+  void bind(@NonNull String name, @NonNull WebviewBindCallback handler);
 
   /**
    * Unbinds a function, removing it from future pages.
    *
    * @param name The name of the function.
    */
-  public void unbind(@NonNull String name) {
-    N.webview_unbind(webview, name);
-  }
+  void unbind(@NonNull String name);
 
   /**
    * Executes an event on the event thread.
    *
    * @implNote Use this only if you absolutely know what you're doing.
    */
-  public void dispatch(@NonNull Runnable handler) {
-    DispatchCallback callback = (wv, arg) -> handler.run();
-
-    // Create upcall stub for the dispatch callback
-    MemorySegment callbackStub =
-        Linker.nativeLinker()
-            .upcallStub(createDispatchCallbackHandle(callback), DispatchCallback.DESCRIPTOR, arena);
-
-    N.webview_dispatch(webview, callbackStub, 0);
-  }
-
-  private static MethodHandle createDispatchCallbackHandle(DispatchCallback callback) {
-    try {
-      return MethodHandles.lookup()
-          .bind(
-              callback,
-              "callback",
-              MethodType.methodType(void.class, MemorySegment.class, long.class));
-    } catch (Exception e) {
-      throw new RuntimeException("Failed to create callback handle", e);
-    }
-  }
+  void dispatch(@NonNull Runnable handler);
 
   /**
    * Executes the webview event loop until the user presses "X" on the window.
@@ -260,54 +106,24 @@ public final class Webview implements Closeable, Runnable {
    * @see #close()
    */
   @Override
-  public void run() {
-    N.webview_run(webview);
-    log.log(DEBUG, "destroy and terminate");
-    N.webview_destroy(webview);
-    N.webview_terminate(webview);
-  }
+  void run();
 
   /**
    * Executes the webview event loop asynchronously until the user presses "X" on the window.
    *
    * @see #close()
    */
-  public void runAsync() {
-    Thread t = new Thread(this);
-    t.setDaemon(false);
-    t.setName("Webview RunAsync Thread - #" + this.hashCode());
-    t.start();
-  }
+  void runAsync();
 
   /** Closes the webview, call this to end the event loop and free up resources. */
   @Override
-  public void close() {
-    log.log(DEBUG, "close");
-    N.webview_terminate(webview);
-  }
+  void close();
 
-  public void setDarkAppearance(boolean shouldAppearDark) {
-    if (WINDOWS.equals(OS_FAMILY)) {
-      WindowsHelper.setWindowAppearance(this, shouldAppearDark);
-    }
-  }
+  void setDarkAppearance(boolean shouldAppearDark);
 
-  public Webview maximizeWindow() {
-    if (WINDOWS.equals(OS_FAMILY)) {
-      WindowsHelper.maximizeWindow(this);
-    }
-    return this;
-  }
+  Webview maximizeWindow();
 
-  public Webview fullscreen() {
-    if (WINDOWS.equals(OS_FAMILY)) {
-      WindowsHelper.fullscreen(this);
-    }
-    return this;
-  }
+  Webview fullscreen();
 
-  public String version() {
-    VersionInfo versionInfo = N.webview_version();
-    return versionInfo.versionNumber();
-  }
+  String version();
 }
