+++ avaje-webview/src/main/java/io/avaje/webview/platform/OSDistribution.java
@@ -1,4 +1,4 @@
-/* 
+/*
 Copyright 2022 Casterlabs
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
@@ -16,88 +16,75 @@ import static io.avaje.webview.platform.OSFamily.UNIX;
 import static io.avaje.webview.platform.OSFamily.VMS;
 import static io.avaje.webview.platform.OSFamily.WINDOWS;
 
-import java.util.regex.Pattern;
+import module java.base;
 
 public enum OSDistribution {
-    // @formatter:off
 
-    // DOS
-    MS_DOS     (DOS,     "MS-DOS",      "MSDOS",   "<manually detected>"),
+  // DOS
+  MS_DOS(DOS, "MS-DOS", "MSDOS", "<manually detected>"),
 
-    // Windows
-    WINDOWS_9X (WINDOWS, "Windows 9x",  "MSDOS",   "windows (95|98|me|ce)"),
-    WINDOWS_NT (WINDOWS, "Windows NT",  "Windows", "win"),
+  // Windows
+  WINDOWS_9X(WINDOWS, "Windows 9x", "MSDOS", "windows (95|98|me|ce)"),
+  WINDOWS_NT(WINDOWS, "Windows NT", "Windows", "win"),
 
-    // Unix
-    MACOS      (UNIX,    "macOS",       "macOS",   "mac|darwin"),
-    SOLARIS    (UNIX,    "Solaris",     "Solaris", "sun|solaris"),
-    BSD        (UNIX,    "BSD",         "BSD",     "bsd"),
-    LINUX      (UNIX,    "Linux",       "Linux",   "nux"),
+  // Unix
+  MACOS(UNIX, "macOS", "macOS", "mac|darwin"),
+  SOLARIS(UNIX, "Solaris", "Solaris", "sun|solaris"),
+  BSD(UNIX, "BSD", "BSD", "bsd"),
+  LINUX(UNIX, "Linux", "Linux", "nux"),
 
-    // VMS
-    OPEN_VMS   (VMS,     "OpenVMS",     "VMS",     "vms"),
+  // VMS
+  OPEN_VMS(VMS, "OpenVMS", "VMS", "vms"),
 
-    /**
-     * This is the fallback, this is not to be considered to be a valid value.
-     */
-    GENERIC    (null,    "Generic",     "Generic", ""),
+  /** This is the fallback, this is not to be considered to be a valid value. */
+  GENERIC(null, "Generic", "Generic", ""),
+  ;
 
-    ;
-    // @formatter:on
+  private final OSFamily family;
 
-    private final OSFamily family;
+  /** A friendly name for the distribution (e.g "macOS" or "Windows NT"). */
+  private final String name;
 
-    /**
-     * A friendly name for the distribution (e.g "macOS" or "Windows NT").
-     */
-    private final String name;
+  private final String regex;
 
-    /**
-     * A "standard" target name.
-     */
-    private final String target;
+  OSDistribution(OSFamily family, String name, String target, String regex) {
+    this.family = family;
+    this.name = name;
+    this.regex = regex;
+  }
 
-    private final String regex;
-
-    OSDistribution(OSFamily family, String name, String target, String regex) {
-        this.family = family;
-        this.name = name;
-        this.target = target;
-        this.regex = regex;
+  static OSDistribution get(OSFamily family) {
+    // If the OS Family is MS DOS then we can't detect it via normal means.
+    // One way is to match path separator which changed in Windows 9x.
+    if ((family == OSFamily.DOS) && ";".equals(System.getProperty("path.separator", ""))) {
+      return MS_DOS;
     }
 
-    static OSDistribution get(OSFamily family) {
-        // If the OS Family is MS DOS then we can't detect it via normal means.
-        // One way is to match path separator which changed in Windows 9x.
-        if ((family == OSFamily.DOS) && System.getProperty("path.separator", "").equals(";")) {
-            return MS_DOS;
-        }
+    String osName = System.getProperty("os.name", "<blank>").toLowerCase();
 
-        String osName = System.getProperty("os.name", "<blank>").toLowerCase();
+    // Loop through the distributions and find one that belongs to the
+    // detected family and matches the regex, returning it if so.
+    for (OSDistribution e : values()) {
+      if (e.family != family) {
+        continue;
+      }
 
-        // Loop through the distributions and find one that belongs to the
-        // detected family and matches the regex, returning it if so.
-        for (OSDistribution e : values()) {
-            if (e.family != family)
-                continue;
-
-            if (Pattern.compile(e.regex).matcher(osName).find()) {
-                return e;
-            }
-        }
-
-        // Fallback.
-        return GENERIC;
+      if (Pattern.compile(e.regex).matcher(osName).find()) {
+        return e;
+      }
     }
 
-    /**
-     * See {@link #name}.
-     *
-     * @return the name of the distribution
-     */
-    @Override
-    public String toString() {
-        return this.name;
-    }
+    // Fallback.
+    return GENERIC;
+  }
 
+  /**
+   * See {@link #name}.
+   *
+   * @return the name of the distribution
+   */
+  @Override
+  public String toString() {
+    return this.name;
+  }
 }
