+++ avaje-webview/src/main/java/io/avaje/webview/platform/OSFamily.java
@@ -1,4 +1,4 @@
-/* 
+/*
 Copyright 2022 Casterlabs
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
@@ -11,54 +11,45 @@ See the License for the specific language governing permissions and limitations 
 */
 package io.avaje.webview.platform;
 
-import java.util.regex.Pattern;
+import module java.base;
 
 public enum OSFamily {
-    // @formatter:off
+  UNIX("Unix", "nux|bsd|.ix|sun|solaris|hp-ux|mac|darwin"),
+  WINDOWS("Windows", "win"),
+  DOS("DOS", "dos"),
+  VMS("VMS", "vms"),
+  GENERIC("Generic", ""),
+  ;
 
-    UNIX     ("Unix",    "nux|bsd|.ix|sun|solaris|hp-ux|mac|darwin"),
-    WINDOWS  ("Windows", "win"),
-    DOS      ("DOS",     "dos"),
-    VMS      ("VMS",     "vms"),
+  /** A friendly name for the family (e.g "Unix" or "Windows"). */
+  public final String name;
 
-    GENERIC  ("Generic", ""),
+  private final String regex;
 
-    ;
-    // @formatter:on
+  OSFamily(String name, String regex) {
+    this.name = name;
+    this.regex = regex;
+  }
 
-    /**
-     * A friendly name for the family (e.g "Unix" or "Windows").
-     */
-    public final String name;
-    private final String regex;
+  static OSFamily get() {
+    String osName = System.getProperty("os.name", "<blank>").toLowerCase();
 
-    OSFamily(String name, String regex) {
-        this.name = name;
-        this.regex = regex;
+    // Search the enums for a match, returning it.
+    for (OSFamily e : values()) {
+      if (Pattern.compile(e.regex).matcher(osName).find()) {
+        return e;
+      }
     }
+    return GENERIC;
+  }
 
-    static OSFamily get() {
-        String osName = System.getProperty("os.name", "<blank>").toLowerCase();
-
-        // Search the enums for a match, returning it.
-        for (OSFamily e : values()) {
-            if (Pattern.compile(e.regex).matcher(osName).find()) {
-                return e;
-            }
-        }
-
-        // Fallback.
-        return GENERIC;
-    }
-
-    /**
-     * See {@link #name}.
-     *
-     * @return the name of the family
-     */
-    @Override
-    public String toString() {
-        return this.name;
-    }
-
+  /**
+   * See {@link #name}.
+   *
+   * @return the name of the family
+   */
+  @Override
+  public String toString() {
+    return this.name;
+  }
 }
