+++ examples/hello-world/src/main/java/example/Main.java
@@ -4,14 +4,11 @@ import io.avaje.webview.Webview;
 
 public class Main {
 
-    static void main() {
+  static void main() {
 
-        Webview webview = Webview.builder()
-                .title("Hi")
-                .html("<h1>Hello World!</h1>")
-                .build();
+    Webview webview = Webview.builder().title("Hi").html("<h1>Hello World!</h1>").build();
 
-        // needs JVM argument -XstartOnFirstThread on Macos
-        webview.run();
-    }
+    // needs JVM argument -XstartOnFirstThread on Macos
+    webview.run();
+  }
 }
