+++ examples/htmx-jex-bootstrap/src/main/java/io/example/data/KBase.java
@@ -2,12 +2,9 @@ package io.example.data;
 
 import java.util.List;
 
-public record KBase(
-        KBaseMeta meta,
-        List<Task> tasks
-) {
-    @Override
-    public String toString() {
-        return meta.name() + " tasks:" + tasks.size();
-    }
+public record KBase(KBaseMeta meta, List<Task> tasks) {
+  @Override
+  public String toString() {
+    return meta.name() + " tasks:" + tasks.size();
+  }
 }
