+++ examples/htmx-jex-bootstrap/src/main/java/io/example/data/KBaseMeta.java
@@ -3,8 +3,4 @@ package io.example.data;
 import io.avaje.jsonb.Json;
 
 @Json
-public record KBaseMeta(
-        String name,
-        String description
-) {
-}
+public record KBaseMeta(String name, String description) {}
