+++ examples/htmx-jex-bootstrap/src/main/java/io/example/data/TaskMeta.java
@@ -4,15 +4,10 @@ import io.avaje.jsonb.Json;
 
 @Json
 public record TaskMeta(
-        String type,
-        String displayName,
-        String searchKeywords,
-        int priority,
-        String description
-) {
-    public String allSearchKeywords() {
-        return searchKeywords == null || searchKeywords.isBlank()
-                ? displayName.toLowerCase()
-                : (displayName + " " + searchKeywords).toLowerCase();
-    }
+    String type, String displayName, String searchKeywords, int priority, String description) {
+  public String allSearchKeywords() {
+    return searchKeywords == null || searchKeywords.isBlank()
+        ? displayName.toLowerCase()
+        : (displayName + " " + searchKeywords).toLowerCase();
+  }
 }
