+++ examples/htmx-jex-bootstrap/src/main/java/io/example/data/package-info.java
