+++ examples/htmx-jex-bootstrap/src/main/java/io/example/web/IndexController.java
@@ -16,23 +16,24 @@ import java.util.List;
 @Controller
 final class IndexController {
 
-    private final DataService dataService;
+  private final DataService dataService;
 
-    IndexController(DataService dataService) {
-        this.dataService = dataService;
-    }
+  IndexController(DataService dataService) {
+    this.dataService = dataService;
+  }
 
-    @Get
-    Page.Index home() {
-        Data data = dataService.data();
-        return new Page.Index(data.kbaseCount(), data.taskCount());
-    }
+  @Get
+  Page.Index home() {
+    Data data = dataService.data();
+    return new Page.Index(data.kbaseCount(), data.taskCount());
+  }
 
-    @HxRequest
-    @Form @Post("search")
-    Partial.SearchTasks searchTasks(String search) {
-        List<Task> tasks = dataService.searchTasks(search, 10);
-        var first = !tasks.isEmpty() ? tasks.getFirst() : null;
-        return new Partial.SearchTasks(first, tasks);
-    }
+  @HxRequest
+  @Form
+  @Post("search")
+  Partial.SearchTasks searchTasks(String search) {
+    List<Task> tasks = dataService.searchTasks(search, 10);
+    var first = !tasks.isEmpty() ? tasks.getFirst() : null;
+    return new Partial.SearchTasks(first, tasks);
+  }
 }
