+++ examples/htmx-jex-bootstrap/src/main/java/io/example/web/StaticConfiguration.java
@@ -9,18 +9,18 @@ import jakarta.inject.Named;
 @Factory
 final class StaticConfiguration {
 
-    @Bean @Named
-    StaticContent favIcon() {
-        return StaticContent.ofClassPath("/static/favicon.ico")
-                .route("/favicon.ico")
-                .build();
-    }
+  @Bean
+  @Named
+  StaticContent favIcon() {
+    return StaticContent.ofClassPath("/static/favicon.ico").route("/favicon.ico").build();
+  }
 
-    @Bean @Named
-    StaticContent staticContent() {
-        return StaticContent.ofClassPath("/static/")
-                .route("/static/*")
-                .directoryIndex("favicon.ico")
-                .build();
-    }
+  @Bean
+  @Named
+  StaticContent staticContent() {
+    return StaticContent.ofClassPath("/static/")
+        .route("/static/*")
+        .directoryIndex("favicon.ico")
+        .build();
+  }
 }
