+++ examples/htmx-jex-bootstrap/src/main/java/io/example/web/view/Page.java
@@ -4,7 +4,6 @@ import io.jstach.jstache.JStache;
 
 public class Page {
 
-    @JStache(path = "index")
-    public record Index(long kbaseCount, long taskCount){}
-
+  @JStache(path = "index")
+  public record Index(long kbaseCount, long taskCount) {}
 }
