+++ examples/htmx-jex-bootstrap/src/main/java/io/example/web/view/package-info.java
